Separated upload page components into their own structure
