+++ frontend/fingertips-frontend/app/upload/page.tsx
@@ -1,4 +1,4 @@
-import { Upload } from '@/components/pages/upload';
+import { Upload } from '@/upload/components/pages/upload';
 
 export default async function UploadPage() {
   return <Upload />;
