+++ frontend/fingertips-frontend/upload/components/organisms/ApiResponsePanel/ApiResponsePanel.styles.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const ResponsePanel = styled('div')({
+  marginBottom: '50px',
+  borderColor: '#1D70B8',
+  borderWidth: '5px',
+  borderStyle: 'solid',
+  padding: '15px',
+});
