+++ frontend/fingertips-frontend/upload/components/pages/upload/Upload.styles.tsx
@@ -1,0 +1,10 @@
+import { H1, WarningText } from 'govuk-react';
+import styled from 'styled-components';
+
+export const InterimWarning = styled(WarningText)({
+  marginBottom: '50px',
+});
+
+export const InterimWarningText = styled('span')({ fontSize: '20px' });
+
+export const PageHeading = styled(H1)({ fontSize: '36px' });
